magit-blame: Introduce magit-blame-file-since
Take a sha1 (or branch name or anything git recognizes) and call
magit-blame-file-on using it. Also check that git-blame is possible
with the given sha1 (this is done with 'git show sha1:file', see the
TODO). A bad sha1 or a renamed/added file will fail this check.

Also, introduce magit-blame-previous-revision which takes the sha1
from the current annotated blame section and calls
magit-blame-file-since with its parent (i.e. sha1^). A helper function
magit-blame-sha1-from-overlay has been refactored out of
magit-blame-locate-commit.